refactor: add tailwind to every component and delete individual css pages
