+++ src/components/Container.jsx
@@ -1,0 +1,3 @@
+export const Container = ({ children, className = "" }) => {
+    return <div className={`max-w-[1290px] mx-auto px-5 ${className}`}>{children}</div>;
+};
