--- src/components/LanguageSelect/LanguageSelect.css
@@ -1,26 +1,0 @@
-.language-select {
-	padding: 4px 6px;
-	border-radius: var(--border-radius);
-	border: 2px solid var(--primary-color);
-	background-color: var(--bg-color);
-	color: var(--text-color);
-	font-family: inherit;
-	font-size: small;
-	cursor: pointer;
-	transition: all 0.2s ease-in-out;
-	appearance: none;
-	-webkit-appearance: none;
-	-moz-appearance: none;
-	position: relative;
-}
-
-.language-select:hover {
-	border-color: var(--primary-color-dark);
-	background-color: var(--section-bg);
-}
-
-.language-select:focus {
-	outline: none;
-	border-color: var(--text-secondary);
-	box-shadow: 0 0 0 2px rgba(18, 100, 163, 0.3);
-}
