--- src/components/LanguageSelect/LanguageSelect.jsx
@@ -1,20 +1,0 @@
-import { useContext } from "react";
-import { LanguagesContext } from "../../context/LanguagesContext";
-import "./LanguageSelect.css";
-
-export const LanguageSelect = () => {
-	const { lang, handleLang } = useContext(LanguagesContext);
-
-	return (
-		<select
-			className="language-select"
-			name="language"
-			id="language"
-			value={lang}
-			onChange={(e) => handleLang(e.target.value)}
-		>
-			<option value="es">🇪🇸 Español</option>
-			<option value="en">🇬🇧 English</option>
-		</select>
-	);
-};
