--- src/components/LoginForm/LoginForm.css
@@ -1,86 +1,0 @@
-.login-form {
-	display: flex;
-	flex-direction: column;
-	gap: var(--spacing-sm);
-	min-width: 200px;
-}
-
-.login-label-input {
-	display: flex;
-	flex-direction: column;
-}
-
-.no-focus:focus {
-	outline: none;
-	box-shadow: 0 0 0 2px rgba(18, 100, 163, 0.3);
-}
-
-.login-label-input input {
-	padding: var(--spacing-sm);
-	border: none;
-	border-radius: var(--border-radius);
-}
-
-.login-label-checked {
-	display: flex;
-	align-items: center;
-	align-self: center;
-}
-
-.login-input-button {
-	display: flex;
-	align-items: center;
-	position: relative;
-}
-
-.login-input-button input {
-	width: 100%;
-}
-
-.eye-btn {
-	display: flex;
-	align-items: center;
-	width: 30px;
-	height: 30px;
-	background: transparent;
-	border: none;
-	padding-right: 10px;
-	margin: 0;
-	position: absolute;
-	right: 0;
-	cursor: pointer;
-}
-
-.register-message {
-	align-self: center;
-}
-
-.registered-link {
-	text-decoration: underline;
-	color: var(--text-secondary);
-}
-
-.submit-btn {
-	font-family: inherit;
-	font-size: inherit;
-	align-self: center;
-	background-color: var(--accent-color);
-	border: 2px solid var(--primary-color);
-	border-radius: var(--border-radius);
-	padding: var(--spacing-sm);
-	margin-top: var(--spacing-sm);
-	width: 80px;
-	box-shadow: var(--button-shadow);
-	transition: all 0.3s ease-in-out;
-}
-
-.submit-btn:hover {
-	background-color: var(--accent-hover);
-	border-color: var(--primary-color-dark);
-	transform: translateY(-2px);
-}
-
-.error-message {
-	align-self: center;
-	color: var(--primary-color);
-}
