+++ src/components/Navbar.jsx
@@ -1,0 +1,1 @@
+export const Navbar = () => {};
