--- src/components/Overlay/Overlay.css
@@ -1,17 +1,0 @@
-.overlay {
-	position: fixed;
-	top: 0;
-	left: 0;
-	width: 100%;
-	height: 100%;
-	background-color: var(--overlay);
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	z-index: 1000;
-}
-
-.overlay-content {
-	max-width: 400px;
-	width: 80%;
-}
