+++ src/index.css
@@ -1,160 +1,152 @@
 @import "tailwindcss";
 
-*,
-*::before,
-*::after {
-	box-sizing: border-box;
+html {
+    position: relative;
+    background: linear-gradient(135deg, var(--bg-primary) 0%, var(--bg-secondary) 100%);
+    background-blend-mode: multiply;
+    overflow: hidden;
 }
 
-html,
-body {
-	font-family: "RoadCaptain", sans-serif;
-	margin: 0;
-	padding: 0;
-}
+html::after {
+    content: "";
+    position: absolute;
+    inset: 0;
+    pointer-events: none;
+    z-index: 0;
 
-/* h1,
-h2,
-h3,
-h4,
-h5,
-h6,
-p,
-ul,
-ol,
-li,
-figure,
-figcaption,
-blockquote,
-dl,
-dd {
-	margin: 0;
-	padding: 0;
-} */
-
-ul {
-	list-style-type: none;
-}
-
-a {
-	color: inherit;
-	text-decoration: none;
+    background-image: repeating-linear-gradient(
+        45deg,
+        rgba(0, 0, 0, 0.015) 0px,
+        rgba(0, 0, 0, 0.015) 2px,
+        transparent 2px,
+        transparent 8px
+    );
+    transform: skewY(-6deg);
 }
 
 @font-face {
-	font-family: "OnePiece";
-	src: url("./assets/fonts/OnePiece.woff2") format("woff2");
-	font-weight: normal;
-	font-style: normal;
+    font-family: "OnePiece";
+    src: url("./assets/fonts/OnePiece.woff2") format("woff2");
+    font-weight: normal;
+    font-style: normal;
 }
 
 @font-face {
-	font-family: "RoadCaptain";
-	src: url("./assets/fonts/RoadCaptain.woff2") format("woff2");
-	font-weight: normal;
-	font-style: normal;
+    font-family: "RoadCaptain";
+    src: url("./assets/fonts/RoadCaptain.woff2") format("woff2");
+    font-weight: normal;
+    font-style: normal;
 }
 
 @font-face {
-	font-family: "GreatVibes";
-	src: url("./assets/fonts/GreatVibes.woff2") format("woff2");
-	font-weight: normal;
-	font-style: normal;
+    font-family: "GreatVibes";
+    src: url("./assets/fonts/GreatVibes.woff2") format("woff2");
+    font-weight: normal;
+    font-style: normal;
 }
 
-:root {
-	/* FONDOS */
-	--bg-color: #f4ede4;
-	--section-bg: #e0d7c9;
-	--card-bg: #fff;
-	--line-color: #414141b0;
-	--color-hover: #d4d6d8;
+@theme {
+    /* COLORES DE FONDO */
+    --color-primary: #f4ede4;
+    --color-secondary: #e0d7c9;
+    --color-card: #fff;
+    --color-default: #414141b0;
+    --color-hover: #d4d6d8;
 
-	/* COLORES DE TEXTO */
-	--text-color: #1d1d1d;
-	--text-secondary: #1264a3;
-	--text-light: #696969;
-	--title-color: #000;
-	--contrast-color: #fff;
+    /* COLORES DE TEXTO */
+    --text-color-primary: #1d1d1d;
+    --text-color-secondary: #1264a3;
+    --text-color-muted: #696969;
+    --text-color-contrast: #fff;
 
-	/* COLORES DE ACCENTO / BOTONES */
-	--accent-color: #ffd700;
-	--accent-hover: #e6b800;
-	--primary-color: #8b0000;
-	--primary-color-dark: #400;
+    /* COLORES DE ACENTO Y BOTONES */
+    --color-accent: #ffd700;
+    --color-accentHover: #e6b800;
+    --color-linePrimary: #8b0000;
+    --color-lineDark: #400;
 
-	/* TIPOGRAFÍAS */
-	--font-title: "OnePiece", serif;
-	--font-main: "RoadCaptain", sans-serif;
-	--font-card: "GreatVibes", cursive;
+    /* FAMILIAS DE FUENTES */
+    --font-family-title: "OnePiece", serif;
+    --font-family-body: "RoadCaptain", sans-serif;
+    --font-family-decorative: "GreatVibes", cursive;
 
-	/* TAMAÑOS DE FUENTE */
-	--font-size-xs: 10px;
-	--font-size-small: 14px;
-	--font-size-base: 16px;
-	--font-size-medium: 18px;
-	--font-size-large: 20px;
-	--font-size-xlarge: 28px;
-	--font-size-xxlarge: 32px;
-	--font-size-title: 70px;
+    /* TAMAÑOS DE FUENTE */
+    --text-xs: 10px;
+    --text-sm: 14px;
+    --text-base: 16px;
+    --text-lg: 18px;
+    --text-xl: 20px;
+    --text-2xl: 28px;
+    --text-3xl: 32px;
+    --text-4xl: 70px;
 
-	/* SPACING */
-	--spacing-xs: 4px;
-	--spacing-sm: 8px;
-	--spacing-md: 16px;
-	--spacing-lg: 24px;
-	--spacing-xl: 32px;
+    /* ESPACIADO */
+    --spacing-1: 4px;
+    --spacing-2: 8px;
+    --spacing-4: 16px;
+    --spacing-6: 24px;
+    --spacing-8: 32px;
 
-	/* CONTENEDORES */
-	--container-width: 95%;
-	--max-container-width: 1200px;
-	--border-radius: 6px;
-	--section-padding: 2rem;
-	--card-padding: 1rem;
+    /* DIMENSIONES DE CONTENEDORES */
+    --width-container: 95%;
+    --max-width-container: 1200px;
+    --rounded: 6px;
+    --padding-section: 2rem;
+    --padding-card: 1rem;
 
-	/* SOMBRAS */
-	--overlay: rgba(0, 0, 0, 0.8);
-	--box-shadow-white: 0px 2px 10px rgba(255, 255, 255, 0.5);
-	--box-shadow: 0px 2px 6px rgba(0, 0, 0, 0.15);
-	--card-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-	--button-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
+    /* SOMBRAS Y EFECTOS */
+    --color-overlay: rgba(0, 0, 0, 0.8);
+    --shadow-white: 0px 0px 20px 2px rgba(255, 255, 255, 0.8);
+    --shadow-default: 0px 2px 6px 1px rgba(0, 0, 0, 0.8);
+    --shadow-card: 0 4px 6px rgba(0, 0, 0, 0.1);
+    --shadow-button: 0 2px 4px rgba(0, 0, 0, 0.64);
 
-	/* TRANSICIONES */
-	--transition-fast: 0.2s ease-in-out;
-	--transition-medium: 0.4s ease-in-out;
+    /* TRANSICIONES */
+    --transition-fast: 0.2s ease-in-out;
+    --transition-default: 0.4s ease-in-out;
 }
 
 .light-theme {
-	/* FONDOS */
-	--bg-color: #ffffff;
-	--section-bg: #f9f9f9;
-	--card-bg: #ffffff;
-	--line-color: #e0e0e0;
-	--color-hover: #d4d6d8;
+    /* FONDOS */
+    --bg-color: #ffffff;
+    --section-bg: #f9f9f9;
+    --card-bg: #ffffff;
+    --line-color: #e0e0e0;
+    --color-hover: #d4d6d8;
 
-	/* COLORES DE TEXTO */
-	--text-color: #222222;
-	--text-secondary: #555555;
-	--text-light: #888888;
-	--title-color: #000000;
-	--contrast-color: #ffffff;
+    /* COLORES DE TEXTO */
+    --text-color: #222222;
+    --text-secondary: #555555;
+    --text-light: #888888;
+    --title-color: #000000;
+    --contrast-color: #ffffff;
 
-	--accent-color: #0073e6;
-	--accent-hover: #005bb5;
-	--primary-color: #400d40;
-	--primary-color-dark: #4a154b;
+    --accent-color: #0073e6;
+    --accent-hover: #005bb5;
+    --primary-color: #400d40;
+    --primary-color-dark: #4a154b;
 
-	/* SOMBRAS */
-	--box-shadow: 0px 2px 6px rgba(0, 0, 0, 0.1);
-	--card-shadow: 0 4px 6px rgba(0, 0, 0, 0.08);
-	--button-shadow: 0 2px 4px rgba(0, 0, 0, 0.15);
+    /* SOMBRAS */
+    --box-shadow: 0px 2px 6px rgba(0, 0, 0, 0.1);
+    --card-shadow: 0 4px 6px rgba(0, 0, 0, 0.08);
+    --button-shadow: 0 2px 4px rgba(0, 0, 0, 0.15);
 }
 
-body {
-	background-color: var(--bg-color);
+.btn {
+    @apply self-center py-1 px-1.5 border-2 min-w-20 border-linePrimary rounded cursor-pointer 
+        transition-all duration-200 ease-in-out appearance-none relative 
+        shadow-button;
 }
 
-.container {
-	width: var(--container-width);
-	margin: 0 auto;
+.btn:hover {
+    @apply -translate-y-0.5 border-lineDark;
+}
+
+.shadow-title {
+    @apply text-shadow-[-2px_-2px_0_#8b0000,_2px_-2px_0_#8b0000,_-2px_2px_0_#8b0000,_2px_2px_0_#8b0000,_0_0_10px_#000];
+}
+
+.no-focus:focus {
+    outline: none;
+    box-shadow: 0 0 0 2px rgba(18, 100, 163, 0.3);
 }
