--- src/layouts/Footer/Footer.css
@@ -1,16 +1,0 @@
-.footer-container {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	text-align: center;
-	gap: var(--spacing-xs);
-	padding: var(--spacing-md) 0;
-	border-top: 1px solid var(--primary-color);
-}
-
-.footer-text {
-	display: flex;
-	flex-direction: column;
-	gap: var(--spacing-xs);
-	font-size: var(--font-size-xs);
-}
