--- src/pages/LoginPage/LoginPage.css
@@ -1,29 +1,0 @@
-.login-container {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	position: relative;
-}
-
-.login-text {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	text-align: center;
-	gap: var(--spacing-xs);
-	padding: var(--spacing-md);
-}
-
-h1 {
-	font-size: var(--font-size-title);
-	color: #ffd700;
-	text-shadow: -2px -2px 0 #8b0000, 2px -2px 0 #8b0000, -2px 2px 0 #8b0000, 2px 2px 0 #8b0000, 0 0 10px #000;
-}
-
-h2 {
-	font-size: var(--font-size-xxlarge);
-	color: #c9a066;
-	text-shadow: 2px 2px 0px #3b2f2f;
-	letter-spacing: 2px;
-	margin-top: -10px;
-}
