--- src/pages/LoginPage/RegisterForm/RegisterForm.css
@@ -1,36 +1,0 @@
-.register-form {
-	display: flex;
-	flex-direction: column;
-	gap: var(--spacing-sm);
-	padding: var(--spacing-sm);
-	background-color: var(--bg-color);
-	border-radius: var(--border-radius);
-	box-shadow: var(--box-shadow-white);
-}
-
-.register-form h3 {
-    align-self: center;
-    font-size: var(--font-size-xlarge);
-}
-
-.register-languages {
-	padding: var(--spacing-sm);
-	border: none;
-	border-radius: var(--border-radius);
-}
-
-.register-checked {
-	display: flex;
-	flex-direction: column;
-	gap: var(--spacing-xs);
-}
-
-.register-checked-terms {
-	display: flex;
-	align-items: center;
-	gap: var(--spacing-xs);
-}
-
-.terms-message {
-	font-size: var(--font-size-xs);
-}
